i broke the code into files but im getting console error
