+++ enumerations.js
@@ -1,0 +1,5 @@
+export const Player = {
+    x: -1,
+    none: 0,
+    o: 1
+}
